Conflicts:
	MessengerMaven/src/main/java/messenger/Implementation/UserManagementImpl.java
	MessengerMaven/src/main/java/messenger/Service/UserServiceImpl.java
	MessengerMaven/src/test/java/messenger/Interface/UserManagementTest.java
	MessengerMaven/target/classes/META-INF/maven/Interface/MessengerMaven/pom.properties
	MessengerMaven/target/classes/messenger/Domain/User.class
	MessengerMaven/target/classes/messenger/Implementation/ManageChatGroupsImpl.class
	MessengerMaven/target/classes/messenger/Implementation/ManageContactListImpl.class
	MessengerMaven/target/classes/messenger/Implementation/UserManagementImpl.class
	MessengerMaven/target/classes/messenger/Interface/UserManagement.class
	MessengerMaven/target/classes/messenger/Service/UserService.class
	MessengerMaven/target/classes/messenger/Service/UserServiceImpl.class

+++ MessengerMaven/src/main/java/messenger/Interface/UserManagement.java
@@ -21,7 +21,7 @@ public interface UserManagement {
 	 * @return true: user deleted successfully<br>
 	 * false: error while deleting
 	 */
-	public boolean deleteUser(int userid);
+	public boolean deleteUser(Long userid);
 	
 	/**
 	 * update user data
@@ -31,7 +31,7 @@ public interface UserManagement {
 	 * @return true: update successful<br>
 	 * false: error while updating
 	 */
-	public boolean updateUser(int userid, String username, String password);
+	public boolean updateUser(Long userid, String username, String password);
 	
 	/**
 	 * checks login credentials
